Added working filters bar
The filter bar works perfectly for the languages, but for the Country and Categories, it sometimes returns 1 or 2 articles only, most times nothing.

I think this is because of the fact that it uses another url, being topheadlines for both of these, and something with the url building is broken, will fix later!